move AbstractHttpServiceIntegrationTestFuns to separate file
CHANGELOG_BEGIN
CHANGELOG_END
